Moved processors out of build
