+++ build.rs
@@ -9,121 +9,15 @@ use std::fs;
 use std::path::PathBuf;
 use std::process::Command;
 use test_generator::{
-    build_ignores_from_textfile, extract_name, test_directory, test_directory_module,
-    with_backends, with_test_module, Test, Testsuite,
+    build_ignores_from_textfile, emscripten_processor, test_directory,
+    test_directory_module, wasi_processor, wast_processor, with_backends, with_test_module,
+    Testsuite,
 };
 
 static EMTESTS_ENV_VAR: &str = "WASM_EMSCRIPTEN_GENERATE_EMTESTS";
 static WASITESTS_ENV_VAR: &str = "WASM_WASI_GENERATE_WASITESTS";
 static WASITESTS_SET_UP_TOOLCHAIN: &str = "WASM_WASI_SET_UP_TOOLCHAIN";
 static WASITESTS_GENERATE_ALL: &str = "WASI_TEST_GENERATE_ALL";
-
-/// Given a Testsuite and a path, process the path in case is a wast
-/// file.
-fn wast_processor(out: &mut Testsuite, p: PathBuf) -> Option<Test> {
-    let ext = p.extension()?;
-    // Only look at wast files.
-    if ext != "wast" {
-        return None;
-    }
-
-    // Ignore files starting with `.`, which could be editor temporary files
-    if p.file_stem()?.to_str()?.starts_with(".") {
-        return None;
-    }
-
-    let testname = extract_name(&p);
-    let body = format!(
-        "crate::run_wast(r#\"{}\"#, \"{}\")",
-        p.display(),
-        out.path.get(0).unwrap()
-    );
-
-    Some(Test {
-        name: testname.to_string(),
-        body: body.to_string(),
-    })
-}
-
-/// Given a Testsuite and a path, process the path in case is a Emscripten
-/// wasm file.
-fn emscripten_processor(out: &mut Testsuite, p: PathBuf) -> Option<Test> {
-    let ext = p.extension()?;
-    // Only look at wast files.
-    if ext != "wasm" {
-        return None;
-    }
-
-    let outfile = {
-        let mut out_ext = p.clone();
-        out_ext.set_extension("out");
-        // let mut txt_ext = p.clone();
-        // txt_ext.set_extension("txt");
-        if out_ext.exists() {
-            out_ext
-        }
-        // else if txt_ext.exists() {
-        //     txt_ext
-        // }
-        else {
-            return None;
-        }
-    };
-
-    let testname = extract_name(&p);
-    let compiler = out.path.get(0).unwrap();
-    let body = format!(
-        "crate::run_emscripten(r#\"{}\"#, r#\"{}\"#, \"{}\")",
-        p.display(),
-        outfile.display(),
-        compiler
-    );
-
-    Some(Test {
-        name: testname.to_string(),
-        body: body.to_string(),
-    })
-}
-
-/// Given a Testsuite and a path, process the path in case is a WASI
-/// wasm file.
-fn wasi_processor(out: &mut Testsuite, p: PathBuf) -> Option<Test> {
-    let ext = p.extension()?;
-    // Only look at wast files.
-    if ext != "wasm" {
-        return None;
-    }
-
-    let outfile = {
-        let mut out_ext = p.clone();
-        out_ext.set_extension("out");
-        // let mut txt_ext = p.clone();
-        // txt_ext.set_extension("txt");
-        if out_ext.exists() {
-            out_ext
-        }
-        // else if txt_ext.exists() {
-        //     txt_ext
-        // }
-        else {
-            return None;
-        }
-    };
-
-    let testname = extract_name(&p);
-    let compiler = out.path.get(0).unwrap();
-    let body = format!(
-        "crate::run_wasi(r#\"{}\"#, r#\"{}\"#, \"{}\")",
-        p.display(),
-        outfile.display(),
-        compiler
-    );
-
-    Some(Test {
-        name: testname.to_string(),
-        body: body.to_string(),
-    })
-}
 
 fn is_truthy_env(name: &str) -> bool {
     env::var(name).map(|n| n == "1").unwrap_or_default()
@@ -224,8 +118,6 @@ fn main() -> anyhow::Result<()> {
 
     let emtests_output = out_dir.join("generated_emtests.rs");
     fs::write(&emtests_output, emtests.buffer)?;
-
-    // println!("WASI: {}", wasitests.buffer);
 
     let wasitests_output = out_dir.join("generated_wasitests.rs");
     fs::write(&wasitests_output, wasitests.buffer)?;
