+++ tests/test-generator/src/lib.rs
@@ -5,7 +5,9 @@
 //!
 //! > This program is inspired/forked from:
 //! > https://github.com/bytecodealliance/wasmtime/blob/master/build.rs
+mod processors;
 
+pub use crate::processors::{emscripten_processor, wasi_processor, wast_processor};
 use anyhow::Context;
 use std::collections::HashSet;
 use std::fmt::Write;
